Separate the common code for a person's social displays and links
